+++ pages/share/get.php
@@ -24,19 +24,22 @@
   </div>
 
   <form role="form" action="/<?= html(SECRET_URI) ?>" method="post">
-    <label for="secret"><h1>Share a Secret:</h1></label>
-    <input type="text" autocomplete="off" class="form-control" id="secret" name="secret" maxlength="512" size="512" required="required" />
+    <h1>Share a Secret:</h1>
+    <div id="secret-div">
+      <textarea autocomplete="off" class="form-control" id="secret" name="secret" rows="5" required="required"></textarea>
+      <div id="counter">1024</div>
+    </div>
     <button type="submit" class="btn btn-default pull-right" id="share-secret-btn" name="share-secret-btn">Share the Secret!</button>
   </form>
-
-  <link href="/resources/css/share.css?<?= $cache_value ?>" integrity="sha256-tByl5f3IGvPqqtUvyHcSIe4SXVXRnx7wiMlmG07yZbA=" rel="stylesheet" type="text/css" />
 
   <label class="checkbox-inline" for="encrypt-locally"><input type="checkbox" autocomplete="off" id="encrypt-locally" value="" />Password-protected: </label>
   <input type="password" autocomplete="off" class="form-control" id="password" maxlength="64" size="32" />
   <input type="button" class="btn btn-default" id="encrypt" value="Protect!" />
 
+  <link href="/resources/css/share.css?<?= $cache_value ?>" integrity="sha256-EYu1Dc10IDi0yUOyV55YWmCKWfVlBj1rTMk/AsbViKE=" rel="stylesheet" type="text/css" />
+
   <script src="/resources/js/lib.js?<?= $cache_value ?>" integrity="sha256-d6lWbau1r7bB+u7utpkOX6tLTXnbOPqD33NmIXHY19A=" type="text/javascript"></script>
-  <script src="/resources/js/share.js?<?= $cache_value ?>" integrity="sha256-olc3XkJFlipNwdHG9nGoIO7VksmEvWYo/K2dnu+YCfk=" type="text/javascript"></script>
+  <script src="/resources/js/share.js?<?= $cache_value ?>" integrity="sha256-jVvsQvz1rwhbs7Ddpihie31jsn/9byFRLI88NnGXzGM=" type="text/javascript"></script>
 
 <?php