+++ resources/js/norepost.js
@@ -1,0 +1,5 @@
+
+  if (window.history.replaceState) {
+    window.history.replaceState(null, null, window.location.href);
+  }
+